+++ libs/cardano-ledger-api/src/Cardano/Ledger/Api/Tx/Wits.hs
@@ -26,6 +26,7 @@ module Cardano.Ledger.Api.Tx.Wits (
   hashDataTxWitsL,
   TxDats (..),
   unTxDats,
+  unTxDatsL,
   rdmrsTxWitsL,
   Redeemers (..),
   unRedeemers,
@@ -56,6 +57,7 @@ import Cardano.Ledger.Alonzo.TxWits (
   unRedeemers,
   unRedeemersL,
   unTxDats,
+  unTxDatsL,
  )
 import Cardano.Ledger.Api.Era ()
 import Cardano.Ledger.Conway.Scripts (ConwayPlutusPurpose (..))